+++ src/provider/aws/neptune/instances/instanceEncrypted.ts
@@ -1,25 +1,26 @@
 import { AuditResultInterface, AWSScannerInterface } from 'cloud-search'
 import { AWS, keyTypeArg } from '../../../../lib/aws/AWS'
-import { DBCluster } from 'aws-sdk/clients/neptune'
+import { DBInstance } from 'aws-sdk/clients/neptune'
 import assert from 'assert'
-const rule = 'ClusterEncrypted'
+const rule = 'InstanceEncrypted'
 
 export const command = `${rule} [args]`
 
-export const desc = `Amazon Neptune graph cluster databases should have encryption at rest enabled
+export const desc = `Amazon Neptune graph database instances should have encryption enabled
 
-  OK      - Neptune cluster encryption is encrypted at rest
-  UNKNOWN - Unable to determine if Neptune cluster encryption enabled
-  FAIL    - Neptune cluster is not encrypted at rest
+  OK      - Neptune instance encryption is encrypted at rest
+  UNKNOWN - Unable to determine if Neptune instance encryption enabled
+  FAIL    - Neptune instance is not encrypted at rest
 
   resourceId: Database Identifier
 
 `
+//TODO: Figure out the resourceId for Neptune.
 export const builder = {
   ...keyTypeArg,
 }
 
-export default class ClusterEncrypted extends AWS {
+export default class InstanceEncrypted extends AWS {
   audits: AuditResultInterface[] = []
   service = 'neptune'
   global = false
@@ -28,18 +29,18 @@ export default class ClusterEncrypted extends AWS {
     super({ ...params, rule })
   }
 
-  async audit({ resource, region }: { resource: DBCluster; region: string }) {
+  async audit({ resource, region }: { resource: DBInstance; region: string }) {
     const options = this.getOptions()
     options.region = region
-    assert(resource.DBClusterIdentifier, 'cluster missing its DB Identifier')
+    assert(resource.DBInstanceIdentifier, 'instance missing its DB Identifier')
     const audit = this.getDefaultAuditObj({
-      resource: resource.DBClusterIdentifier,
+      resource: resource.DBInstanceIdentifier,
       region,
     })
     if (resource.KmsKeyId) {
       assert(
         this.keyType,
-        'Key type is required argument for isKeyTrusted check'
+        'Key type is required arguement for isKeyTrusted check'
       )
       audit.state = await this.isKeyTrusted(
         resource.KmsKeyId,
@@ -59,30 +60,30 @@ export default class ClusterEncrypted extends AWS {
     resourceId: string
     region: string
   }) => {
-    let clusters
+    let instances
     const options = this.getOptions()
     options.region = region
     if (resourceId) {
       const promise = new this.AWS.Neptune(options)
-        .describeDBClusters({
-          DBClusterIdentifier: resourceId,
+        .describeDBInstances({
+          DBInstanceIdentifier: resourceId,
         })
         .promise()
-      clusters = await this.pager<DBCluster>(promise, 'DBClusters')
+      instances = await this.pager<DBInstance>(promise, 'DBInstances')
     } else {
       const promise = new this.AWS.Neptune(options)
-        .describeDBClusters()
+        .describeDBInstances()
         .promise()
-      clusters = await this.pager<DBCluster>(promise, 'DBClusters')
+      instances = await this.pager<DBInstance>(promise, 'DBInstances')
     }
-    for (const cluster of clusters) {
-      await this.audit({ resource: cluster, region })
+    for (const instance of instances) {
+      await this.audit({ resource: instance, region })
     }
   }
 }
 
 export const handler = async (args: AWSScannerInterface) => {
-  const scanner = new ClusterEncrypted(args)
+  const scanner = new InstanceEncrypted(args)
   await scanner.start()
   scanner.output()
   return scanner.audits