+++ src/provider/aws/neptune/neptune.ts
@@ -10,11 +10,11 @@ const opts = {
 export const builder = (yargs: Argv) => {
   return yargs
     .commandDir('clusters', opts)
-    // .commandDir('instances', opts)
+    .commandDir('instances', opts)
     .demandCommand(1)
     .example([
       [
-        `neptune clusters <rule>...`,
+        'neptune clusters <rule>...',
         'have the cli show you available cluster rule',
       ],
     ])